refactored out the interaction to the api via a "wrapper" class
